feat: 경매 등록 기능 구현 (#9)
* feat : Auction, Image, Product, User Repository 인터페이스 추가
(#30)

* feat : Product Entity 내 Category Enum 내용 추가
(#30)

* refactor : confing -> config 디렉토리 이름 오타 수정
(#30)

* feat : Auction Entity에 minPrice의 Custom Validation 구현
(#30)

* feat : Auction Entity에 Status Enum 구체화
(#30)

* feat : Custom ExceptionHandler 구현 및 UserNotFoundException 추가
(#30)

* refactor : GeneratedValue 전략 수정 및 minPrice 타입 int로 통일
(#30)

* feat : 경매 등록 요청 DTO 추가
(#30)

* feat : AuctionService 서비스 추가
(#30)

* feat : 경매 등록 API 추가
(#30)

* feat : Image Service 서비스 추가
- ImageUploader 테스트 용 인터페이스 추가
- 상품 이미지 업로드
- S3, CloudFront 저장 로직 구현
(#30)

* chore : AWS 환경 세팅
- AWS Java SDK 의존성 추가
- AWS Config S3 클라이언트 연동
- AWS 보안 Key 추가
(#30)

* refactor : Image Entity @Entity 어노테이션 추가
(#30)

* test : 경매 등록 테스트 코드 추가
- test profile 내 Spring Security CSRF 비활성화, 모든 요청 허용
(#30)

* refactor : 경매 물건 이미지 저장 클래스 수정
- 53 이미지 등록 기능 메소드 별도 클래스로 분리
- 잘린 cdn 주소 DB에 저장하는 메소드 리턴 값 List로 변경
(#30

* test : 경매 등록 테스트 및 CI 빌드 성공
(#30

* refactor: 경매 등록 응답 값 수정
- 응답 본문 없이 상태 코드와 Location 헤더만 반환
(#30

* refactor: 경매 엔티티 컬럼 타입 수정
- minPrice 컬럼 int -> Integer 수정 반영
(#30

* refactor: 이미지 등록 관련 로그 수정
- println -> slf4j 로 변경
(#30

* refactor: toList() 메서드 반영
(#30

* refactor: @Transactional(readOnly = true) 반영

* refactor: S3 Bucket Bean 이름 등록 처리

* refactor: Category Enum 어노테이션 적용
- Getter
- AllArgsConstructor
(#30)

* refactor: MVC 테스트 환경 재구축
- SecurityConfigTest.java 삭제 후 로컬 세팅
(#30)

* refactor: 경매 등록 테스트 로컬 세팅
(#30)

* feat: 이미지 Custom 에러 처리 구현
(#30)

* refactor: AWSConfig s3BucketName 메소드 수정
- S3BucketName -> s3BucketName 네이밍 컨벤션
- Bean 등록하지 않고 필요할 때 가져와서 사용
(#30)

* refactor: Custom Exception 추가 수정
(#30)

* refactor: 경매 등록 서비스 최종 병합
(#30)

* refactor: Custom Exception 추가 수정
- ImageUploadException, UserNotFoundException 추가
- getCode 메소드 추가
(#30)

* refactor: generateFileName 메소드 제거
(#30)

* refactor: AuctionControllerTest 세팅
- Spring Security Juit Test에 적용
(#30)

* refactor: application-test.yml 파일 추가
- CI 테스트 통과 세팅
(#30)

* refactor: CI 환경변수 최종 테스트
(#30)

* refactor: CI 환경변수 최종 테스트 2
(#30)

* refactor: CI 환경변수 최종 테스트 3
(#30)

* refactor: CI 환경변수 최종 테스트 4
- secret -> secrets
- CI env 코드 제외
- @SpringBootTest 활성화
(#30)

* refactor: CI 환경변수 최종 테스트 5
- build.gradle 'profile active' 코드 추가
- CI 코드에 환경 변수 직접 명시
(#30)

* refactor: CI 환경변수 최종 테스트 6
- spring -> cloud 변경
(#30)

* refactor: CI 환경변수 최종 테스트 7
- '\' 줄바꿈 명령어 추가
(#30)

* refactor: 경매 등록 테스트 Mock 이미지 구현
- @SpringBootTest 주석 처리 한 부분 해결 필요
(#30)

* refactor: 경매 등록 테스트 S3 의존성 문제 해결
- S3 모의 객체 Test 전역 설정
- TestAWSConfig -> AWSConfig 클래스명 변경
- AWSConfig Profile -> local, prod 로 제한
(#30)

* feat: Custom Exception 기본 코드 수정
(#30)

* feat: 이미지 비즈니스 예외 추가
(#30)

* feat: 유저 비즈니스 예외 추가
(#30)

* feat: 유저 비즈니스 에러코드 추가
- 유저를 찾을 수 없을 때 발생하는 에러 코드
(#30)

* feat: 이미지 비즈니스 에러코드 추가
- 이미지 저장에 실패했을 때 발생하는 에러코드
(#30)

* feat: 사용자 데이터 조회 시 에러코드 반영
(#30)

* feat: 이미지 저장 시 에러코드 반영
(#30)

+++ src/main/java/org/chzz/market/domain/auction/dto/AuctionResponse.java
@@ -16,12 +16,12 @@ public class AuctionResponse {
     private String name;
     private String cdnPath;
     private Long timeRemaining;
-    private Long minPrice;
+    private Integer minPrice;
     private Long participantCount;
     private Boolean isParticipating;
 
     @QueryProjection
-    public AuctionResponse(Long id, String name, String cdnPath, LocalDateTime createdAt, Long minPrice,
+    public AuctionResponse(Long id, String name, String cdnPath, LocalDateTime createdAt, Integer minPrice,
                            Long participantCount,
                            Boolean isParticipating) {
         this.id = id;
